+++ src/modules/edc-demo/components/asset-viewer/asset-viewer.component.ts
@@ -1,9 +1,10 @@
 import {Component, OnInit} from '@angular/core';
 import {MatDialog} from '@angular/material/dialog';
-import {BehaviorSubject, Observable, of} from 'rxjs';
+import {BehaviorSubject} from 'rxjs';
 import {map, switchMap} from 'rxjs/operators';
 import {AssetService} from '../../../edc-dmgmt-client';
 import {Asset} from '../../models/asset';
+import {Fetched} from '../../models/fetched';
 import {AssetPropertyMapper} from '../../services/asset-property-mapper';
 import {AssetDetailDialogData} from '../asset-detail-dialog/asset-detail-dialog-data';
 import {AssetDetailDialogResult} from '../asset-detail-dialog/asset-detail-dialog-result';
@@ -11,13 +12,18 @@ import {AssetDetailDialog} from '../asset-detail-dialog/asset-detail-dialog.comp
 import {AssetEditorDialogResult} from '../asset-editor-dialog/asset-editor-dialog-result';
 import {AssetEditorDialog} from '../asset-editor-dialog/asset-editor-dialog.component';
 
+export interface AssetList {
+  filteredAssets: Asset[];
+  numTotalAssets: number;
+}
+
 @Component({
   selector: 'edc-demo-asset-viewer',
   templateUrl: './asset-viewer.component.html',
   styleUrls: ['./asset-viewer.component.scss'],
 })
 export class AssetViewerComponent implements OnInit {
-  filteredAssets$: Observable<Asset[]> = of([]);
+  assetList: Fetched<AssetList> = Fetched.empty();
   searchText = '';
   private fetch$ = new BehaviorSubject(null);
 
@@ -28,29 +34,27 @@ export class AssetViewerComponent implements OnInit {
   ) {}
 
   ngOnInit(): void {
-    this.filteredAssets$ = this.fetch$.pipe(
-      switchMap(() => {
-        let assets$ = this.assetService
-          .getAllAssets()
-          .pipe(
-            map((assets) =>
-              assets.map((asset) =>
-                this.assetPropertyMapper.readProperties(asset.properties),
-              ),
+    this.fetch$
+      .pipe(
+        switchMap(() => {
+          return this.assetService.getAllAssets().pipe(
+            map(
+              (assets): AssetList => ({
+                filteredAssets: assets
+                  .map((asset) =>
+                    this.assetPropertyMapper.readProperties(asset.properties),
+                  )
+                  .filter((asset) => asset.name?.includes(this.searchText)),
+                numTotalAssets: assets.length,
+              }),
             ),
+            Fetched.wrap({
+              failureMessage: 'Failed fetching asset list.',
+            }),
           );
-
-        if (this.searchText) {
-          assets$ = assets$.pipe(
-            map((assets) =>
-              assets.filter((asset) => asset.name?.includes(this.searchText)),
-            ),
-          );
-        }
-
-        return assets$;
-      }),
-    );
+        }),
+      )
+      .subscribe((assetList) => (this.assetList = assetList));
   }
 
   onSearch() {