+++ src/modules/edc-demo/components/catalog-browser/catalog-browser.component.ts
@@ -1,14 +1,19 @@
 import {Component, OnDestroy, OnInit} from '@angular/core';
 import {MatDialog} from '@angular/material/dialog';
-import {BehaviorSubject, Subject, sampleTime} from 'rxjs';
+import {BehaviorSubject, Observable, Subject, sampleTime} from 'rxjs';
 import {map, switchMap, takeUntil} from 'rxjs/operators';
-import {Asset} from '../../models/asset';
 import {ContractOffer} from '../../models/contract-offer';
+import {Fetched} from '../../models/fetched';
 import {CatalogApiUrlService} from '../../services/catalog-api-url.service';
 import {CatalogBrowserService} from '../../services/catalog-browser.service';
 import {AssetDetailDialogData} from '../asset-detail-dialog/asset-detail-dialog-data';
 import {AssetDetailDialogResult} from '../asset-detail-dialog/asset-detail-dialog-result';
 import {AssetDetailDialog} from '../asset-detail-dialog/asset-detail-dialog.component';
+
+export interface ContractOfferList {
+  filteredContractOffers: ContractOffer[];
+  numTotalContractOffer: number;
+}
 
 @Component({
   selector: 'edc-demo-catalog-browser',
@@ -16,9 +21,7 @@ import {AssetDetailDialog} from '../asset-detail-dialog/asset-detail-dialog.comp
   styleUrls: ['./catalog-browser.component.scss'],
 })
 export class CatalogBrowserComponent implements OnInit, OnDestroy {
-  filteredContractOffers: ContractOffer[] = [];
-  filteredContractOfferAssets: Asset[] = [];
-
+  contractOffersList: Fetched<ContractOfferList> = Fetched.empty();
   searchText = '';
   customProviders = '';
   presetProvidersMessage = '';
@@ -34,20 +37,29 @@ export class CatalogBrowserComponent implements OnInit, OnDestroy {
     this.fetch$
       .pipe(
         sampleTime(200),
-        switchMap(() => this.catalogBrowserService.getContractOffers()),
-        map((contractOffers) => {
-          const searchText = this.searchText.toLowerCase();
-          return contractOffers.filter((contractOffer) =>
-            contractOffer.asset.name?.toLowerCase()?.includes(searchText),
-          );
-        }),
+        switchMap(
+          (): Observable<Fetched<ContractOfferList>> =>
+            this.catalogBrowserService.getContractOffers().pipe(
+              map(
+                (contractOffers): ContractOfferList => ({
+                  filteredContractOffers: contractOffers.filter(
+                    (contractOffer) =>
+                      contractOffer.asset.name
+                        ?.toLowerCase()
+                        ?.includes(this.searchText.toLowerCase()),
+                  ),
+                  numTotalContractOffer: contractOffers.length,
+                }),
+              ),
+              Fetched.wrap({
+                failureMessage: 'Failed fetching contract offers.',
+              }),
+            ),
+        ),
         takeUntil(this.ngOnDestroy$),
       )
-      .subscribe((filteredContractOffers) => {
-        this.filteredContractOffers = filteredContractOffers;
-        this.filteredContractOfferAssets = filteredContractOffers.map(
-          (it) => it.asset,
-        );
+      .subscribe((contractOffersList) => {
+        this.contractOffersList = contractOffersList;
       });
     this.presetProvidersMessage = this.buildPresetCatalogUrlsMessage();
   }